JRuby String#bytes returns enum
Calling table.bytes.size is failing in JRuby because of an
[incompatability between JRuby and MRI][1].

In MRI 1.8 and 1.9 String#bytes returned an Enumerator. MRI 2.x changes
this behavior to return an array. JRuby <= 1.7.x mimics the 1.9
behavior of returning an Enumerator.

The simple solution is to call to_a on the returned bytes when we are
checking the byte size of the newly assigned table.

  [1]: https://github.com/jruby/jruby/issues/1109